FenceSync now creates and uses InsInfo.

+++ src/machine.cpp
@@ -1183,7 +1183,7 @@ void Machine::test(){
           return false;
         }
         for(unsigned i = 0; i < fb.size(); ++i){
-          if(fb[i] < 0 || fb[i] >= m->automata[i].get_states().size()){
+          if(fb[i] < 0 || fb[i] >= (int)m->automata[i].get_states().size()){
             return false;
           }
         }
@@ -1216,7 +1216,7 @@ void Machine::test(){
           return false;
         }
         for(unsigned i = 0; i < fb.size(); ++i){
-          if(fb[i] < 0 || fb[i] >= m->automata[i].get_states().size()){
+          if(fb[i] < 0 || fb[i] >= (int)m->automata[i].get_states().size()){
             return false;
           }
         }
@@ -1249,7 +1249,7 @@ void Machine::test(){
           return false;
         }
         for(unsigned i = 0; i < fb.size(); ++i){
-          if(fb[i] < 0 || fb[i] >= m->automata[i].get_states().size()){
+          if(fb[i] < 0 || fb[i] >= (int)m->automata[i].get_states().size()){
             return false;
           }
         }
@@ -1284,7 +1284,7 @@ void Machine::test(){
           return false;
         }
         for(unsigned i = 0; i < fb.size(); ++i){
-          if(fb[i] < 0 || fb[i] >= m->automata[i].get_states().size()){
+          if(fb[i] < 0 || fb[i] >= (int)m->automata[i].get_states().size()){
             return false;
           }
         }
